Embeded offset/fetch with the sort functionality
